# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AIGC Pool API
  version: 0.1.0
  summary: 多模型 AIGC 创作平台的对外契约（用户端 + 管理端）
  description: |
    本文件是 **前后端唯一契约**，前端据此 mock 并行开发，后端据此实现。

    ## 与 capability-schema.ts 的关系

    `#/components/schemas/ModelCapabilitySchema` 及其所有子 schema 是
    `docs/contracts/capability-schema.ts` 的逐字段镜像。**两者必须同步修改**，
    任一方单独改动即视为破坏契约。TypeScript 侧是权威定义，本文件是它的传输层表述。

    ## 约定

    - 所有时间戳为 RFC3339（UTC，例 `2026-08-07T15:04:05Z`）。
    - 所有 ID 为字符串，前端**只做透传，不解析其内容**。
    - 所有错误响应统一为 `#/components/schemas/ErrorEnvelope`，前端按 `error.code` 分支，
      **不解析 `error.message`**。
    - 列表接口一律**游标分页**（`cursor` / `next_cursor`），不提供 offset。
      理由：资产瀑布流在有新产物时 offset 会重复/漏项。
    - 认证为 `Authorization: Bearer <JWT>`。管理接口额外要求 `role=admin`，
      路由前缀 `/api/admin`，越权返回 403 `forbidden`。

    ## 隐式分流

    `POST /api/tasks` **没有 mode 字段**。传了 `inputs.reference_images` 就是图生图，
    没传就是文生图，由后端按输入槽是否为空路由。前端不参与这个判断。
servers:
  - url: http://localhost:18080
    description: 本地开发（本阶段不部署）
...
security:
  - bearerAuth: []
paths:
...
  /api/tasks/{taskId}/cancel:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [tasks]
      summary: 取消任务
      operationId: cancelTask
      description: |
        `queued` 可取消并全额退回冻结积分；`running` 尽力取消（能取消的上游调用上游取消，
        不能的标记为 canceled 并**不扣费**）；终态任务返回 409。
      responses:
        '200':
          description: 已取消
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '409':
          description: 任务已处于终态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  ############################################################
  # stream
  ############################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    JsonValue:
      description: 任意 JSON 值（对应 capability-schema.ts 的 JsonValue）
    Modality:
      type: string
      enum: [image, video]
    Timestamp:
      type: string
      format: date-time
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ApiError' }
      description: |
        **所有**错误响应统一为该结构。前端严格按 `error.code` 呈现，不解析 message。
    TaskErrorCode:
      type: string
      description: |
        **与 capability-schema.ts 的 `TaskErrorCode` 逐值一致，不多不少。**
        前三类是 PRD 定义的失败三分类，**三类都不扣费**
        （`charged` 恒为 false，但仍显式回传，前端据此显示「未扣费」）。
      enum:
        - invalid_param
        - upstream_rate_limited
        - content_rejected
        - insufficient_credit
        - internal_error
    ApiErrorCode:
      type: string
      description: |
        HTTP 层错误码 = `TaskErrorCode` 全集 + 若干传输层码。
        分成两个枚举而不是往 `TaskErrorCode` 里塞，是为了让 `TaskErrorCode`
        与 TypeScript 契约保持**逐值可对拍**：任务失败分类是产品语义，
        401/403/404 是传输语义，混在一起前端的失败三分类分支会被污染。
        前端遇到未知 code 按 `internal_error` 降级呈现，不允许白屏。
      enum:
        - invalid_param
        - upstream_rate_limited
        - content_rejected
        - insufficient_credit
        - internal_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - revision_conflict
        - rate_limited
        - quota_exceeded
    ApiError:
      allOf:
        - $ref: '#/components/schemas/TaskError'
        - type: object
          properties:
            code: { $ref: '#/components/schemas/ApiErrorCode' }
    TaskError:
      type: object
      description: |
        `docs/contracts/capability-schema.ts` 的 `TaskError` 的镜像。
        挂在 `Task.error` 与 SSE `task.failed` 上。
      required: [code, message, retryable, charged]
      properties:
        code: { $ref: '#/components/schemas/TaskErrorCode' }
        message: { type: string }
        field_errors:
          type: array
          description: 仅 invalid_param 携带，key 对应 ParamSpec.key 或 InputSlotSpec.key
          items:
            type: object
            required: [key, message]
            properties:
              key: { type: string }
              message: { type: string }
        retryable: { type: boolean }
        charged:
          type: boolean
          description: 失败三分类恒为 false
        retry:
          type: object
          description: upstream_rate_limited 时携带，用于展示「正在自动重试 (2/5)」
          required: [attempt, max_attempts]
          properties:
            attempt: { type: integer }
            max_attempts: { type: integer }
            next_at: { $ref: '#/components/schemas/Timestamp' }

    ####################################################
    # auth / me
    ####################################################
...
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
      description: |
        **平台自己的状态枚举。** 各上游供应商的状态在 L0 driver 内部归一化到这里，
        L2 以上只认平台状态（见 tech-design §状态归一化表）。
    Task:
      type: object
      required: [id, model_id, modality, status, prompt, params, created_at]
      properties:
        id: { type: string }
        model_id: { type: string }
        model_name: { type: string }
        modality: { $ref: '#/components/schemas/Modality' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        progress:
          type: [number, 'null']
          minimum: 0
          maximum: 1
          description: 供应商大多不给真实百分比，前端靠 eta 做乐观推进（上限卡 90%）
        queue_position: { type: [integer, 'null'] }
        eta_seconds: { type: [number, 'null'] }
        prompt: { type: string }
        params:
          type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
        inputs:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        estimated_cost: { type: integer }
        actual_cost:
          type: [integer, 'null']
          description: 仅成功后有值；失败三分类恒为 0
        assets:
          type: array
          items: { $ref: '#/components/schemas/Asset' }
        error:
          oneOf:
            - $ref: '#/components/schemas/TaskError'
            - type: 'null'
        canvas_id: { type: [string, 'null'] }
        card_id: { type: [string, 'null'] }
        client_token: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
...
    AssetType:
      type: string
      enum: [image, video, audio, text]
    Asset:
      type: object
      description: |
        **必须带多档 URL**（前端按画布缩放分级加载，否则同屏 20+ 个原图/video 必炸），
        **必须带 source**（否则「做同款」和「单卡重跑」同时废掉）。
        所有 URL 均指向本平台存储，绝不是上游 URL。
      required: [id, type, original, created_at]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/AssetType' }
        original:
          type: string
          description: 原图/原视频地址
        thumb_512:
          type: [string, 'null']
          description: 512px 缩略图；画布 0.4 ≤ k < 1 时用
        poster:
          type: [string, 'null']
          description: 视频封面；`<video>` 未挂载时显示
        mime: { type: string }
        bytes: { type: integer, format: int64 }
        width: { type: [integer, 'null'] }
        height: { type: [integer, 'null'] }
        duration_ms: { type: [integer, 'null'] }
        task_id: { type: [string, 'null'] }
        source: { $ref: '#/components/schemas/AssetSource' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
    AssetSource:
      type: object
      description: 生成参数快照。血缘的「怎么来的」那一半，与 lineage 的「从谁来的」互补。
      required: [model_id, prompt, params, input_asset_ids]
      properties:
        model_id: { type: string }
        prompt: { type: string }
        params:
          type: object
          additionalProperties: { $ref: '#/components/schemas/JsonValue' }
        input_asset_ids:
          type: array
          items: { type: string }
        skill_id: { type: [string, 'null'] }
...